Basic O/R mapping using Hibernate for topics and users with
- a set association from topic to a set of users,
  the authors that have contributed to the topic

- a map association from topics to topics denoting
  the subtopics of a topic indexed by name

This entails that looking up a topic /Foo/Bar/Baz
requires looking up the topics / /Foo /Foo/Bar and
finally /Foo/Bar/Baz. Topics themselves are not
identified by name, but by their identifier. This
makes it possible to link to the same topic from
different places in the hierarchy.

Imported the classes from the user1/ experiment
to build the webinterface.


svn path=/; revision=80