M    TODO
ideas, ideas, ideas

M    src/org/webdsl/dsl/syntax/WebDSL.sdf

sketch of HQL syntax; requires a more thorough approach

M    src/org/webdsl/dsl/transformation/desugar.str

fixed bug in editing of associations; edits of Sets did not work
because changes were made to the derived List property, which are not
propagated to the original Set. Fixed by defining a version of the
InputAssociation desugaring rule for properties of Set type, which
applies the additions/removals to the set rather than the list; while
the dataTable is (for) is still based on the list.

svn path=/; revision=282