Store diffs of pages on edits. Using diff_match_patch library of Neil Fraser to compute and apply patches between strings.
http://code.google.com/p/google-diff-match-patch/

Added Patch type to store patches as text, with operations makePatch
and applyPatch. Library should eventually be added as a jar somewhere,
(We should consider the binding of language constructs to the run-time
system) but is now added as Java files. The patch_factory is my
addition to the library, making the diff_match_patch object a
singleton.


svn path=/; revision=572